국가별 import market의 price를 int -> Long으로 바꿈

+++ web_project/src/main/java/net/kdigital/web_project/dto/BycounImMarketDTO.java
@@ -20,9 +20,8 @@ public class BycounImMarketDTO {
 	private int dateYear;
 	private int ranking;
 	private String importMarket;
-	private int price;
+	private Long price;
 	private String percentile;
-	
 	public static BycounImMarketDTO toDTO(BycounImMarketEntity bycounImMarketEntity) {
 		return BycounImMarketDTO.builder()
 				.seq5(bycounImMarketEntity.getSeq5())
